fix: don't create dependencies between rewriting and better-auth
